Fixed HIA.a bug

+++ src/be/ac/kuleuven/cs/drama/vertalerpack/vertaler/Vertaler2.java
@@ -74,7 +74,7 @@ public class Vertaler2 {
 
    public int translateLine(int line) {
       try {
-         return (Integer) _linemap.get(line - 1);
+         return ((Integer) _linemap.get(line - 1)).intValue();
       } catch (Exception e) {
          System.out.println("*** " + e);
          return -1;
@@ -190,7 +190,7 @@ public class Vertaler2 {
 
    /**
     * @param symbol the symbol name
-    * @result the numeric value of the symbol
+    * @return  the numeric value of the symbol
     * @exception AbnormalTerminationException the symbol is not defined
     */
    public int getSymbolValue(String symbol)
